+++ tests/integrations/notdiamond/custom_router_test.py
@@ -9,6 +9,7 @@ import pytest
 import weave
 from weave.flow.eval import EvaluationResults
 from weave.integrations.notdiamond.custom_router import train_evaluations, evaluate_router
+from weave.integrations.notdiamond.util import get_model_evals
 from weave.trace.weave_client import WeaveClient
 
 @pytest.fixture
@@ -17,9 +18,6 @@ def model_evals():
 
 @pytest.fixture
 def model_datasets(model_evals: Dict[str, EvaluationResults]):
-    return get_model_datasets(model_evals)
-
-def get_model_datasets(model_evals: Dict[str, EvaluationResults]):
     model_datasets = {}
     for model, eval_results in model_evals.items():
 
@@ -46,39 +44,6 @@ def preference_id():
 
     response_body = cassette['interactions'][0]['response']['body']
     return json.loads(response_body['string'])['preference_id']
-
-def get_model_evals(
-    file_path: str = "integrations/notdiamond/test_data/humaneval.csv",
-) -> Dict[str, EvaluationResults]:
-    df = pd.read_csv(file_path)
-    models = [
-        "anthropic/claude-3-5-sonnet-20240620",
-        "openai/gpt-4o-2024-05-13",
-        "google/gemini-1.5-pro-latest",
-        "openai/gpt-4-turbo-2024-04-09",
-        "anthropic/claude-3-opus-20240229",
-    ]
-
-    model_evals = {}
-    for model in models:
-        input_score_col = _get_score_column(model)
-        input_response_col = _get_response_column(model)
-        columns = ["Input", input_response_col, input_score_col]
-        eval_rows = (
-            df[columns]
-            .rename(columns={"Input": "prompt", input_response_col: "actual", input_score_col: "scores"})
-        )
-        eval_rows['scores'] = eval_rows['scores'].apply(lambda x: {"correctness": x})
-        eval_rows = eval_rows.to_dict(orient="records")
-        model_evals[model] = EvaluationResults(rows=weave.Table(eval_rows))
-
-    return model_evals
-
-def _get_score_column(model: str) -> str:
-    return f"{model}/final_score"
-
-def _get_response_column(model: str) -> str:
-    return f"{model}/response"
 
 @pytest.mark.skip_clickhouse_client
 @pytest.mark.vcr(
